pyproject.toml
[project]
name = "recruitment-support-fbref"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "soccerdata>=1.8.0",
  "pandas>=2.2",
  "numpy>=2.0",
  "pyarrow>=16.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.5",
  "typer>=0.12",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
